Implement initialConfig(), test fixture for RswMhnic class
Summary:
Now that we have a top-level function to generate prod-setting MHNIC configs, we call this function from within HwProdInvariantsRswMhnicTest's `initialConfig()`.

Along with this, we also define helper functions lifted from HwQueuePerHostRouteTests to be called in the setup phase of RswMhnic's test fixture.

Miscellaneously, I forgot to set RswTest and FswTest's invariant options to `DEFAULT_INVARIANTS`. This is done here.

Reviewed By: jasmeetbagga

Differential Revision: D30108894

fbshipit-source-id: 1bc72f27cd84a53d46fae32c315e296dbee6a164