Function to get all downlink ports from SwitchConfig
Summary:
For dynamic config testing, we will need to be able to separate the downlinks from uplinks on RSWs in the case that we are testing a queue-per-host device. On RswMhnic's, olympic QOS is applied to uplinks and QPH is applied to downlinks; this function will allow us to distinguish which ports we're testing on.

The function works by checking each of the config's vlanPorts, and returning all PortID's for which the PortID is equal to `kDownlinkBaseVlanId`.

Differential Revision: D30120940

fbshipit-source-id: 9a44e3c663a541befe2a5c2291baf00864f395aa