Introduce localMac as new input of Platform::Platform()
Summary:
As discussed in D30272445, adding localMac as a new input for Platform constructor.

I was thinking to hide this local mac address inside PlatformProductInfo so that we don't have to touch all existing Platforms.
But since PlatformProductInfo is only used for a product info for a Platform, which is learned from netwhoami or fruid.json.
And Local Mac is actually read from eth0 or defined by some mock mac address for testing,
we shouldn't mix these two together.

Therefore, I went with an extra input for all the constructors everywhere.

Now for prod hardware, we will fetch the local mac address once in:
- WedgePlatformInit::createWedgePlatform() for BCM SDK platforms
- SaiPlatformInit::initSaiPlatform() for SAI SDK platforms

Reviewed By: jasmeetbagga

Differential Revision: D30351541

fbshipit-source-id: 24dd667a272a73898a2c0013cc397d01607ea65e