reproducing cont canary issue in sai
Summary:
as titled. this is done in response to cont canary issue. we made some changes in our copp acls and got rid of two higher priority acls which were no longer needed due to ip2me routes. as a result priority of all acls shift up.

one of these acls is cpuPolicing-CPU-Port-Mcast-v6-ff02::/16

this acl is special, it has source port as one of the qualifier and permits traffic from cpu instead of trapping back.

what makes this acl special, is that qualifiers in this acl (src port) is not supported when it is set to "cpu port". further this acl has an action "permit" which is also currently not supported currently.

because these are not supported, acl sai entry is not created and placed in internal map. this means entry for "cpuPolicing-CPU-Port-Mcast-v6-ff02::/16" does not exist.

in response to config change (acl pri change), the fboss attempts to change this acl, which includes removing this acl and readding it back. since this acl does not in map removal fails with an exception

```
C++ exception with description "attempted to remove aclEntry which does not exist: cpuPolicing-CPU-Port-Mcast-v6-ff02::/16" thrown in the test body.
```

same exception is seen in cont canary

TODO:

fix for this in next diff

Reviewed By: shri-khare

Differential Revision: D27549038

fbshipit-source-id: b2dc0703ff8c7c292e51a54f9fb2cfbe544dba6c