Refactor to get uplink or downlink ports from SwitchConfig
Summary:
For QoS verification on MHNIC devices, we need to be able to extract both uplink (olympic) and downlink (queue-per-host) ports from a given SwitchConfig.

To get uplinks, we refactor the getRswDownlinks function from the previous diff into a generalized collectRswPorts that can return both uplinks and downlinks.

Note that this does not work on RTSWs, where downlink vlans start at 2001.

Differential Revision: D30347958

fbshipit-source-id: f46d10bc2953fe0a9a62fbc49d68df417a3b361c