Define functions to get uplinks and downlinks from any platform
Summary:
Now we define general functions to get uplinks and downlinks from arbitrary platforms, not just RSWs.

These functions call the previously-defined getRswUp/Downlinks if the platform provided is an RSW; otherwise it considers the first 4 ports (or the given value of kEcmpWidth) in the config's master list to be uplinks and the rest to be downlinks.

Differential Revision: D30354591

fbshipit-source-id: 0336ba330724880e783952d48233a78c523a31bd