T98755995: ODS counter qsfp_service.qsfp.port.eth1/0/1.present shows 0 when transceiver 0 is present
Summary:
When there is at least 1 transceiver (other than transceiver 0) not present, the ODS counter qsfp_service.qsfp.port.eth1/0/1.present shows 0 when transceiver 0 is present. This is a SW bug in qsfp_servicce. Note the corresponding new format of the same ODS counter, i.e. qsfp_service.qsfp.interface.eth1/1/1.present on wedge100, 400 or qsfp_service.qsfp.interface.eth2/1/1.present on yam, minipack, shows a correct value of 1 when transceiver 0 is present.

The issue is StatsPublisher::publishFbagentCounters() uses *info.port_ref() to get transceiver ID, but WedgeManager::getTransceiversInfo() leaves port_ref() at its default value 0 and only sets present_ref() to false when a transceiver is not present. So when transceiver 0 is present, and there is at least 1 other non-present transceiver, WedgeManager::getTransceiversInfo() will wrongly overwrite present_ref() of transceiver 0 to false when it encounters a non-present transceiver in its for loop. A fix in this diff is to assign a proper transceiver ID to port_ref() in this case.

Differential Revision: D30465180

fbshipit-source-id: 560f58f6def04d2eff02d23f23d05d9f4b9f5643