Use u64HBO to get u64 for intf MAC
Summary:
As titled. We use u64 representation of intfMac to offset
and compute src MAC address. At this point the computation is on
the host itself. So no need to convert into NBO. The above bug
was benign when we were dealing with artificial MAC addresses
in HwTest, viz 02:0:0:0:0:1. But with real MAC addresses,
- link, mnode tests - there is a real risk of creating a bogus
MAC addresses. For e.g. with NBO, the MAC: 9a:5d:82:9:3a:d9 produced
srcMAC of 09:82:5d:9a:00:01. This is a multicast MAC. Packets from this
MAC get dropped on some chips.

Differential Revision: D30477312

fbshipit-source-id: a8f86ce3cfc8bb5264338dc0997123b879de27fe