Refactor isLoadBalanced and programRoutes to accept specific ports rather than ECMP width
Summary:
As it stands now, the top-level functions `isLoadBalanced()` and `programRoutes()` take an ECMP width int as an argument, from which that many random ports are chosen to have ECMP set up on.

To be able to pass specific ports, such as those extracted from a config, we introduce overloads of these functions that take in specific PortDescriptors and use them to call the underlying functions. This will be particularly useful in HwProdInvariants, so that we can setup ECMP and verify load balancing with specific ports (i.e. uplinks).

Reviewed By: jasmeetbagga

Differential Revision: D30432909

fbshipit-source-id: 4844ae92f00483a969a97912e8dc8e1ce836045d