Temporary workaround for async logger to unblock link testing
Summary:
As titled. Sai Replayer was built as a singleton to handle only one instance of Sai Replayer. However qsfp service now creates multiple instances of Sai switch in the same process. The workaround was only to avoid data race.

Before rebuilding Async Logger and Sai Replayer, we'll need to discuss more on the use case of Sai Replayer on Elbert and understand the requirements.

Differential Revision: D30457404

fbshipit-source-id: 3bd304205c0187879bff8f6821818fb954873e37