Towards allowing for trapping packets from multiple ports/ips
Summary:
As titled. Next diffs will allow for creating entries
for multiple IPs/ports. Initially I thought of stacking HwPacketTrapEntry
(say by holding a vector of HwPacketTrapEntry). But this does not
work as in Sai HwTestPacketTrapEntry always creates a entry with
priority of 1, assuming its the only one active. This works if all
the interesting ports, ips are passed in

Reviewed By: zechengh09

Differential Revision: D30533592

fbshipit-source-id: 1c83145a775b7bbb55895bf0195f030abf228c90