Introduce a generic MediaInterfaceCode
Summary:
As discussed offline w/ hgulati, it looks like current qsfp_service is using a slightly complicated union type `MediaInterfaceUnion`
https://www.internalfb.com/code/fbsource/[01c66bef4ff9]/fbcode/fboss/qsfp_service/if/transceiver.thrift?lines=189-192
Which if the module is CMIS, we will use smfCode while if it's SFF, we will use extendedSpecificationComplianceCode.

The original design is to make it easier for SFFModule or CMISModule to read their register values and convert to its own enum directly.
However, this leaves the users to handle the complicated logic to find the correct media interface code, like fboss cli to print this field from TransceiverInfo.

Therefore, we decided to introduce a generic MediaInterfaceCode to represent all possible media interface codes regardless of module types.
And let each single QsfpModule derived class(SFFModule or CMISModule) to convert the raw code to this new enum type.

Reviewed By: birdsoup

Differential Revision: D30653252

fbshipit-source-id: 00f0ef7c089c9e086d077228165eb5cd7382681e