Introduce state/Transceiver.h
Summary:
As we discussed in https://docs.google.com/document/d/1ziltsSmuaQIcHz6hDUJa_TFWZe-X8oIoql4N2bqvpc8/edit#heading=h.3wi8ufae5dtj,
we are moving forward to the approach that have qsfp_service to drive the whole port programming.

To do so, we need wedge_agent to provide a new programInternalPhyPorts(TransceiverInfo, bool force) thrift call to program iphy port inside
wedge_agent. And since qsfp_service is the source of truth of TransceiverInfo and control the lifetime of a transceiver, we are thinking to
deprecate QsfpCache to call qsfp_service thrift getTransceiverInfo().

Instead we use this programInternalPhyPorts to build the new Transceiver object in SwitchState. And having this new state object to program
all the other state objects.

In this diff, we introduce the new state/Transceiver object.

Differential Revision: D30503349

fbshipit-source-id: 5d97a172f8e23cb38b312510f915dd358ad990b7