Require qsfpConfig for qsfp_service
Summary: As titled. By now, coop should be generating qsfp_config on all the prod boxes. For link/qsfp/multi-node tests, the test will generate the qsfp_configs. So it's better to require qsfp config now otherwise it will be hard to debug issues where a module isn't configured according to the config

Reviewed By: jasmeetbagga

Differential Revision: D30172085

fbshipit-source-id: 32e332397b6ac25c2ae43b566787227c9f50ccbe