Add snapshots after link flap
Summary:
Add method in SnapshotManager to post the next X snapshots that are collected. We can then call this when a link flaps to capture snapshots after link flap events.

Also change the syncPorts logic to trigger snapshots only on link state change, not on first port sync (don't publish lots of snapshots on every qsfp service restart)

also reduce cached snapshots to 10 so that we post 10 before and 10 after the link flap event

Differential Revision: D30582707

fbshipit-source-id: 5ecf546958ccc33bedfea8f2c353d690fec7a515