Don't consider programming RxEqualizerSettings based on the status of unused lanes
Summary:
as titled. Currently We will try to program settings on lanes 1-4 again, even if the unused lanes 5-8 need to be programmed. If lanes 5-8 are unused, we shouldn't even consider them in figuring out if the settings need to be programmed or not.
Reprogramming settings on already programmed lanes triggers a link flap on certain modules

Reviewed By: rajank7

Differential Revision: D30687205

fbshipit-source-id: 22a51e97d3725e5a26f62817f9b2aca34de88196