Move all code for initing and querying APIs to SaiApiTable
Summary:
As titled. Have code in one place rather than spreading it around
SaiSwitch and SaiApiTable. Besides better localization, this is needed for
the case where we have more than one SaiSwitch. This is the case in
case of xphys. We don;t want to needlessly keep calling api
init for each SaiSwitch instance.

Differential Revision: D30681194

fbshipit-source-id: 0954d7d72cda6717a8e90bdf6f7a7a89a28538a7