Reset thrift server in shutdown path
Summary:
Observing a crash in thrift server with traceback P454163348 while a graceful shutdown of process is happening. The crash is happening in the message receive path due to a SEGV. As part of the shutdown, we call stopListening() but as per thrift team stopListening() does not stop received message processing if server type is duplex.

Fix:
Stop the server completely in the shutdown path. As per thrift team, duplex mode is depreciated and should be replaced with streaming service. This is a bigger change that we need to do in near future.

Differential Revision: D30659821

fbshipit-source-id: c1f4cd28b60c124318e984b983aea964f54153e6