Updating BMC Client
Summary:
Encountered two issues for improvement with BMCClient.  First issue is inability to reuse BMC client for multiple URL calls.

Second issue is traceback when a problem with the return is handled.

Moved the CurlClient creation into the fetchRaw scope so that a new object gets created each time.

Added try/catch logic to fetchRaw as well to return an empty JSON string if the HTTP call fails.  This is much easier to handle with the multitude of platforms and variety of endpoints we have in the field.

Differential Revision: D30681791

fbshipit-source-id: 378d08f668d8048a2c4a0c2efe0e2c4b70e07769