waitForQsfpService init in link tests
Summary:
We need qsfp svc to be be up and inited before we start programming
ports. This is particularly required for elbert where phy inits from qsfp
svc can take a while.
Right now the code is only for link tests, but wondering if we should add
the same check for prod agent too. One downside of that would be that we
will start fate sharing with qsfp svc. So if qsfp svc is crashing we won't
be able to restart agent either.

Reviewed By: rajank7

Differential Revision: D30769063

fbshipit-source-id: 3ed0bd74e4b781149e0c3b4b1b2d5e5e860cbfe4