Private moduleStateMachine_ in QsfpModule
Summary:
Now with all the existing ModuleStateMachine events have they corresponding ModuleStateMachineEvent enum.
And all the process_event calls are gated by QsfpModule::stateUpdate() function.
We can make the moduleStateMachine_ member as a private member in QsfpModule.

This way will make it easier for us to deprecate the old StateMachine in QsfpModule and use the new StateMachine in TransceiverManager later.

Differential Revision: D30719638

fbshipit-source-id: 65fcc55712829b7c801f19ab99deecfb99fac3ff