Introduce AclTableGroupMap
Summary:
This is towards supporting Multiple ACL Table in FBOSS.

Current use case requires single AclTableGroup (ingress with switch binding).
Thus, switch_config only supports that.

However, it is easier (and cleaner) to implement this in SwSwitch (and consume
in HwSwitch) as a AclTableGroupMap. It will only hold one entry today, but can
be extended later.

Reviewed By: jasmeetbagga

Differential Revision: D30762819

fbshipit-source-id: 120ed169606998a9a6636088db848e5edd0f8114