AclStage as member and ID for ACL Table Group
Summary:
This is towards supporting Multiple ACL Table in FBOSS.

As titled.
AclStage uniquely identifies an ACL Table Group.
We could probably remove the 'name' altogether or keep it for readability.
This diff leaves name as is, as it was already there, but uses stage as the ID.

This is further useful for later stacked diffs that add AclTableGroupMap and
use Stage as the key.

Reviewed By: jasmeetbagga

Differential Revision: D30762818

fbshipit-source-id: 5d535c3c1ae515f3fd47a90dadba79a69809aff8