Correct the addMacsec logic for ingress and egress macsec
Summary: There was an issue. in addMacsec logic where we create the macsec handle for a given direction. Problem is that when we try to create the macsec handle for ingress then we create egress handle also with bypass enable. Then when we try to create the egress handle then we are not turning off the bypass enable. This is causing the bypass to remain enabled even after Tx and Rx macsec is programmed. This is corrected in this diff.

Reviewed By: jasmeetbagga

Differential Revision: D30771139

fbshipit-source-id: 9f7be9762cb3bb0a3943f98a487f10ea02d0b10a