Fix mka session update bug.
Summary:
2 bugs fixed here
- When we get a identical MKAConfig in provisionCAK call, the
desired behavior is noop. However current code was creating a MKASession
w/o checking for a match. MKASession in its destructor cause packet
transport to be deleted. Breaking the existing session in turn. Fix that.
- On an actual update of CAK the same behavior would occur. Since we are
using map emplace, nothing would get inserted in the map. But the temp
MKASession object would break packet transport on its way out. So erase
existing entry before emplacing a new one.

There is actually a third bug, which I suspect will happen when CAK
changes. In that MKASession is a shared pointer, which gets passed on
to pkt io thread. So the MKASession delete may not happen when we
erase from the map. A later MKASession destruction would end up
breaking transport for a updated session on that port. Will fix
in a later diff

Reviewed By: g-shankaran

Differential Revision: D30853578

fbshipit-source-id: 1760db76325f50c0fcfff50af73a7b15d992b37a