Added a helper function to retry refreshTransceivers on unsuccessful attempts
Summary:
There is a lot of flakiness in QSFP hw tests because refresh fails sometimes due to failed I2C transactions. This is not a problem in prod because the next refresh cycles can be successful bringing back the desired behavior. However, the tests today rely on a single refresh to successfully complete which doesn't always happen.

This change adds a helper function to retry refreshTransceivers up to 3 times. refreshTransceivers is deemed successful when the timestamp in TransceiverInfo advances.

Reviewed By: jasmeetbagga

Differential Revision: D30890598

fbshipit-source-id: 41806bf492e4dcfae6064e6a73884fc11a8ec905