compareSerdesAttributes for Tajo
Summary:
Compare only certain attributs for Tajo. Skip idriver attribute since we do not set the attribute whereas an empty list is returned by Tajo.

I will be following up with another diff suggested by phshaikh with a cleaner approach but creating this to unblock on the tx settings sev.

Reviewed By: zechengh09

Differential Revision: D30893928

fbshipit-source-id: 65f0801feee815d40fd00ccee1c8f5c558f111f5