Give LinkSnapshots their own event type
Summary:
We want to separate the linksnapshots from FBOSS_EVENTs for the purpose of rate-limiting. Each event type has its own hourly rate limit, which we were hitting frequently with the link snapshots sharing the same limit as the FBOSS_EVENTs.

Will update the config to post these to scuba

Differential Revision: D30913824

fbshipit-source-id: e7af8ed953160e15b2c77cbf012488f561492fa6