Convert all of AclEntry and MatchAction to Thrifty
Summary:
This is the first part of a big effort to convert the entire switch state to thrift. The motivation behind this is to be able to easily transmit agent's state over thrift while maintaining type safety. This will then be used in fsdb to maintain oper state (see D28019247). An added bonus will be will be we'll hopefully have a less bugs caused by non-backward compatible changes made to agent state manual serialization/deserialization

I'm using Thrifty here which was introduced in way back in D6996910 (https://github.com/facebook/fboss/commit/826fce663df74bd3a055f709ccae8eab6dec6913). The idea is to first represent the state as both cpp classes and an identical thrift struct. Then instead of converting to folly::dynamic manually, we convert to our thrift struct and and Thrifty will handle converting to/from json. This is already being done for Port state.

The biggest challenge will be to keep all this backwards compatible. There are a few cases I've run into which need some sort of custom solution:
1. enums - There are many cases where we're encoding enums using the name and decoding by parsing the string. We can continue to do that in thrift by representing the enum as a string but this practice has caused issues in the past due crashes or inability to downgrade (S201970). So preferably I'd like to change to actually storing an enum if possible
2. Cpp objects not following a nested structure - This is the case where objects in a Fields class are unwrapped in the json encoding then wrapped again during decode. i.e `A { B { int x=1, y=2 }` is encoded as `{"a": { "Bx": 1, "By": 2}}` instead of the thrift encoding which would look like `{"a" : { "b": {"x": 1, "y": 2}}}` (see MatchAction.sendToQueue)
3. Renamed objects - Mainly cosmetic/style

Here I'm trying to solve this by creating a migration step which works forwards and backwards. The migration can be removed after a disruptive push.

Reviewed By: aeckert

Differential Revision: D30532734

fbshipit-source-id: 831e07f19377f2915f83286e5e9117ee5537f239