Fix the thread safe issue b/w disable and get port statCollection
Summary:
Due to enable/disable statCollection and get statCollection are currently
called by different threads, and enable/disable statCollection might need
to attach/detach FlexCounter, which means if we're trying to collect stats
while it's in the process of detaching FlexCounter, the updateStats will
fail. Therefore, we need to make sure these calls from different threads
can maintain exclusive.
Besides since we always reset `portStats_` to nullopt when we disable
the statsCollection, we don't actually need an extra bool to represent
whether the statCollection is enabled or not. What's more, since
 `portStats_` is already Synchornized, we can just utilize it to make sure
enable/disable and get stats are thread safe.

More detailed debug info is in T100029331

Reviewed By: jasmeetbagga

Differential Revision: D30815945

fbshipit-source-id: e65f868471a4e47d16bdc8d0b2362a040b1d11aa