ignore a non-existent entry while removing an acl entry
Summary:
this can happen for acl entry which uses source port as qualifier. both config and sw switch state indicates this entry exists but isn't created on platforms like wedge400c.

to fix this properly, a configuration for wedge400c shouldn't have this acl entry. but not having this acl entry triggers removal of non-existent entry exception. to sanitize wedge400c configuration which doesn't have this entry and to bring config and sw state same as SDK state, introducing this temporary change.

Reviewed By: msomasundaran

Differential Revision: D31028756

fbshipit-source-id: 65d6e98ce1924aaadfe5a4c0a0f5481647b66a21