Route scale test medium term scale projections
Summary:
We anticipate a need for going upto 100K prefixes in the midterm.
This is up from ~50K prefixes scale on FA UUs. The test creates a
100K route scale distribution by
a) Doubling the FA UU route scale distribution. So each subnet scale
is doubled. For e.g. if we had X /56 for UU scale, we have 2X /56
prefixes in this distribution
b) Add 4K /64. The above gets us to ~96K prefixes. Add 4K v6 prefixes
to get us to slightly over 100K prefixes

Reviewed By: msomasundaran

Differential Revision: D31010120

fbshipit-source-id: fcd359789badc7fe040543fe59194d8313efa343