Add support for an optimal ucmp weight distribution method
Summary:
Problem:
Older platforms such as trident2 and tomahawk has support for only 64 or 128 ecmp members. With ucmp, the total weight exceeds the limit quickly and we need to normalize the weights to fit width. The existing algorithm tries to fit the ucmps by rounding up or down and fitting it to the exact width. This does not provide optimal solution because the algorithm does not measure deviation and reduce it. A case study of this problem is available at https://fb.quip.com/QKwQAs1h4HU0

Adding a new distribution algorithm which measures deviation and reduces the maximum error deviation by using a greedy approach. The details are in the code comments. The new algorithm is currently under a flag so that we can test this in few switches before enabling it by default.

The old algorithm has o(n) complexity where n is the number of members. The new algorithm has o(n)*ecmp_width complexity.

Reviewed By: saifhhasan

Differential Revision: D30545807

fbshipit-source-id: c4361a3abdbe254b6937a7076170078bc1ecfc68