Skip checking iDriver and Preemphasis if not set
Summary: As titled. On Sai TH3 we don't program IDriver and Preemphasis, hence the default values are zero in SaiStore. However, across warmboot Brcm-sai would give us some other default values, causing fboss to reprogram the serdes and thus causing link flaps. This is a temporary workaround to skip preemphasis and idriver if we're not setting them.

Reviewed By: jasmeetbagga

Differential Revision: D31069626

fbshipit-source-id: c94c58809a799da1b1ea848823df842ef388f642