Initialize from config file with command line flags
Summary:
To actually enable loading a SwitchConfig dynamically, we introduce the command line flag `--dynamic_config` to explicitly say that we want to load the config from the `--config` argument.

The choice was between explicitly triggering either dynamic loading or static loading. Since on-diff tests don't use a prod config, and we can easily modify Netcastle's commands to run with an extra flag, we decided to explicitly trigger dynamic loading.

Reviewed By: jasmeetbagga

Differential Revision: D30692360

fbshipit-source-id: 37bddec5dc61b06ef11d617d35c0aea961882e04