Bring back forward declaration of SaiHwPlatform
Summary:
Forward declaring SaiHwPlatform requires us to define
and move destructor to .cpp file. Now since we have a user
defined destructor, the compiler does not generate move
constructor for us. So we need to explicitly default it.

Separate diff for ease of review. This is more about
c++ nuance than what the code intends to do

Reviewed By: g-shankaran

Differential Revision: D31236783

fbshipit-source-id: 282c9047d30e771d02ec79fb18a4e5e715c6fed8