Specialize changePort for phy, npu SaiSwitch
Summary:
Similar to addPort, changePort also needs specialization
for phy layer. Previously we were only calling addPort from SaiPhyManager.
Now with the move to processing StateDelta in phy layer, we will
need to handle changePort as well. For now just make this
a pass through to addPort in phy layer (since that's the
only function implemented).

Reviewed By: joseph5wu

Differential Revision: D31242683

fbshipit-source-id: 80ff42cc85579fa06d074cdb34a3428053ce5bb6