Prepare profileConfig and pinConfigs when creating SW ports
Summary:
During agent starts(warmboot and coldboot), we will try to create a PortMap based on either the current
hw state(in coldboot) or warmboot cache.
This is similar as how we prepare the profileID for sw port.
```
    swPort->setProfileConfig(
        *platformPort->getPortProfileConfig(swPort->getProfileID()).iphy_ref());
    swPort->resetPinConfigs(
        platformPort->getIphyPinConfigs(swPort->getProfileID()));
```
The above functions are currently used when BcmPort/SaiPortManager tries to program the HwPort.

So it's safe to use for warmboot as the new change will just make the current logic to call these two functions twice.

The first time is when it tries to build the whole SwitchState from warmboot cache.
But because these two new fields are recently added,
the old wedge_agent version might not cache them in the warmboot cache file.
By manually setting them when we are trying to restore the init state, it will make sure
when we try to program the init config later, these two fields will still be the same.
Therefore, it won't trigger a hardware reprogramming.

Reviewed By: jasmeetbagga

Differential Revision: D31105509

fbshipit-source-id: 3ca3c3949a4ae3c5b65e12f972276270c9ab317d