Allow 256 width wide ECMP
Summary:
Wide ECMP is assumed to be 512 wide in current code. We have a use case now for 256 wide ECMP.

This change still restricts the max to 512. This is because there are instances in code such as during ecmp shrink and expand where we need to assume a max size to read from sdk. If there is a use case in future for larger size, then we can pay the penality

Reviewed By: saifhhasan

Differential Revision: D31410104

fbshipit-source-id: 77747be3c66e0c7bdbe6f20c7cb64f00ef2b9902