Support updating Port attributes in programInternalPhyPorts
Summary:
Previous diff already supports changing Port profileConfig and pinConfigs in ApplyThriftConfig.cpp which is used for loading config.
As we also support ThriftHandler::programInternalPhyPorts() to allow qsfp_service call wedge_agent to update Transceiver map,
we also need to update the port profileConfig and pinConfig with the new state of Transceiver.

Reviewed By: jasmeetbagga

Differential Revision: D31282130

fbshipit-source-id: d4c5b7559145724cea18620be7bb38508ef5444f