Support Wedge100 customize override factor
Summary:
Currently Wedge100 is using cable length as a factor to pick different iphy serdes settings for downlink profiles.
And such override has a range of supported cable lengths.
https://www.internalfb.com/code/configerator/[704df2cdd4de]/source/neteng/fboss/platform_mapping/wedge100.cinc?lines=55-106

Therefore, if the cable length is longer than 3m, we need to use the 3m iphy serdes settings; whilc if cable length is short than 1m, using 1m settings.
To do so, we need to customize the override factor for both Wedge100Platform and SaiBcmWedge100Platform.
Otherwise we might skip using the right settings to program the serdes.

Reviewed By: jasmeetbagga

Differential Revision: D31250039

fbshipit-source-id: 10a33585d54622bd34b9fbf1c11407b749d1facd