Use QsfpCache to update TransceiverMap in SwitchState when applying configs
Summary:
As the goal is to remove the dependency of TransceiverInfo in BcmPort and SaiPortManager when programming port profile config and pin configs.
Previous diffs already support TransceiverMap in SwitchState and preparing ProfileConfig and PinConfigs in software ports.
Now we need to update the TransceiverMap in SwitchState with QsfpCache before updatePorts()
so that we can use the current TransceiverInfo to build the right ProfileConfig and PinConfigs for hw to program ports.

Reviewed By: jasmeetbagga

Differential Revision: D31212923

fbshipit-source-id: 6addb2d6e6825445d968c387e3c32639050f5c3d