Introduce new ModuleStateMachine
Summary:
The  new module state machine is planned to replace
the current `moduleStateMachine` and `modulePortStateMachine`. Because the
change is quite massive and right now they have been heavily used in
QsfpModule, it might be safer to start the new change on a new state machine
rather than tamper the current one.

Reviewed By: birdsoup

Differential Revision: D30746532

fbshipit-source-id: 9539c0692b3cab55edd28d8a8c2222c3d429e802