Fix for Rx SAK getting deleted right after creation
Summary: The Rx SAK was getting deleted right after programming it. The reason is that the function SaiPortManager::programMacsec() keeps a list called oldRxSaks having existing Rx SAK. If the new SAK does not exist or does not match with old SAK then it creates new SAK . But when new SAK already exists and it is same as existing SAK then it goes on deleting the old SAK which is a problem. The fix is to prevent deletion of old SAK if the new SAK is same as old SAK and we are not re-programming it.

Reviewed By: jasmeetbagga

Differential Revision: D31466185

fbshipit-source-id: 30cc19faa786c6aa4cf739408b97d62778bb964a