Create ThriftPath plugin
Summary:
Instead of hand-creating the path types, this moves
generating path types in to a thrift plugin. This makes it possible to
use the fsdb libs to work with arbitrary thrift types.

In this diff, I moved our service + some tests to wrap the AgentConfig
thrift object, as opposed to the test thrift type we had been using.

Differential Revision: D29995584

fbshipit-source-id: deb116b17ead84e8a8528fc5a7a2377bb809ae63