Add support for getPlatform()->getPhyInterfaceHandler()
Summary: Want to be able to do getPlatform()->getPhyInterfaceHandler() without crazy stunts. This seems to be needed by other code as well. They currently cast the platform into specific platform. Due to the cast to different platforms, it needs copies of the same code for Minipack, Elbert, etc. The interface is common to all platforms anyways.

Reviewed By: birdsoup

Differential Revision: D31589797

fbshipit-source-id: 7c49ebe84cb477eb652482fb9a089727ed9b2df5