Subcommand args [3/7] - Add helpers to turn arguments into typed tuple
Summary:
Theoretically if we have a tuple type with the args in order, at runtime getTypedArgs can create a tuple with the correct size and values. We need two tuple types here, one with std::monostates filtered out and one with them in there. This will be used to figure out which positional args we need. As an example, say we have a command:

`fboss2 show A <args1> B C <args2>`

here `B` does not take any arguments, so the final args should look like std::tuple<Args1, Args2>. getTypedArgs will first create a tuple of size 3, then filter out the second position base on the `Types` argument

Differential Revision: D31462722

fbshipit-source-id: 423b00d91f16fe76d2e2c24ba1078fc6f40c4171