Subcommand args [1/7] - Create CmdArgsLists
Summary:
This is the first diff in the stack to support nested subcommand args. As such I will explain the overall solution here.

# Problem
Need functionality to have subcommand args at arbitrary levels. For example something like `fboss2 show bgp neighbours <sender_peer-addr> some-subcommand <recevier_peer-addr>`. In the current structure we store arguments by object type and call queryClient with a single argument depending on the argument type. This won't scale since we could have multiple different positional arg lists of the same type, and queryClient will need a varying number of params depending on which subcommand

# Goal
My ultimate goal with this implementation was to have as much of the code in the framework as possible, leaving trivial amount of argument handling in the subcommand impl. Specifically, I wanted to have the subcommand be able to define:

`RetType queryClient(HostInfo hostInfo, Arg1T arg1, Arg2T arg2, ..., ArgNT argN) { ... }`

The alternative to this was to have queryClient take some object like this CmdArgsLists, and find it's arguments, but that wouldn't be ideal since now we'll need some logic in subcommands to handle arguments.

# Solution
The outline of the solution is as follows:

1. Firstly, every subcommand will need to define
    a. The object type it needs at the that level. i.e CmdShowPortQueue will need a queueId at that level
    b. The parent command. i.e. CmdShowPort is parent of CmdShowPortQueue
2. Since every command will define it's parent (if there is any), we will essentially have a linked list at compile time. We can use this linked list to recursively create a tuple type for all the arguments of a command. I make two versions of this, 1 with the monostates (no arg subcommands) filtered away, and one with them there. This will be used at runtime to determine which positions we need
`ArgTypes = arg_type_of<Subcommand> = tuple<arg_type_of<Subcommand::Parent>..., Subcommand::ArgT>`
3. At runtime we throw all arguments into an `ArgsArray = array<vector<string>, MAX_DEPTH>`, where MAX_DEPTH is the theoretical maximum depth subcommand we can have (we can always increase this, we just need something to make the compiler happy)
4. Now we can turn our `ArgsArray` into an `ArgsTuple = tuple<vector<string> * N>`, (here im denoting a tuple of size N = depth of subcommand, where every value is a string vector).
6. Then we filter out positions in the tuple that are not needed (using a filter on std::monostate which denotes no arg subcommands)
5. OPTIONAL: Right now all object types in subcommands are of type vector<string>, but optionally they can be a class that has an implicit constructor `MyCustomArgClass(vector<string> args) {...}`, and we can implicitly turn our vector tuple to a fully typed tuple with a class for each object type
6. Now that we have a fully typed ArgsTuple, we can call queryClient using `std::apply(queryClient, ArgsTuple)`

So now queryClient can define its arguments based on it's parents and it's own arguments.

For reference, initial design proposal for this was in P459937611. Implementation details have changed since then but the general idea is the same.

Differential Revision: D31355974

fbshipit-source-id: 6016ff84c2e2f30ed8f8f83cba5d8af48847cb17