Adding a new 200g-fr4 module to firmware upgrader
Summary: Added support for the new Innolight 200G FR4 module which has a different part number. Currently there was a 1:1 map in the code for module name to part number. But now we understood that this map needs to be 1:N because for a module type (ie innolight-200g-fr4) there could be more than one part number. The diff addresses it

Differential Revision: D31725176

fbshipit-source-id: d8b1a98046f870f2ba0e15144a2f72939c3eb862