Introduce triggerProgrammingEvent() for new state machine logic
Summary:
Introduced triggerProgrammingEvent() in WedgeManager::refreshTransceivers().
This will check whether the new state machine has passed the iphy/ xphy/ transceiver programmed state.
If not, it will trigger on port programming.

The reason I put this new function in WedgeManager::refreshTransceivers() is to take advantage of current periodic
full transceivers refresh so that if the port programming failed, it will pick up again in next refresh.

Reviewed By: birdsoup

Differential Revision: D31810893

fbshipit-source-id: cd2eecc0c4aca8d84b6c38c3ad004ebb882d4ff4