New StateMachine: call wedge_agent programInternalPhyPorts()
Summary:
Introduce `TransceiverManager::programInternalPhyPorts(TransceiverID id)` so that the new state machine
can start programming the iphy ports.
The reason I stored a TransceiverManager pointer as an attribute in the state machine and have TransceiverManager to handle the main
programming logic, instead of having StateMachine to do so is
I believe the state machine should just be the tool to help us understand the state and action of a transceiver but keep the complexity of
functionality to something like TransceiverManager.

Reviewed By: birdsoup

Differential Revision: D31810894

fbshipit-source-id: addd8b3030b0419d40cbeaf58be6a7441a565e0b